:warning: move declarative plugin to golang domain
