+++ pkg/plugins/golang/declarative/v1/internal/templates/controller.go
