+++ pkg/plugins/golang/domain.go
@@ -1,12 +1,9 @@
 /*
 Copyright 2020 The Kubernetes Authors.
-
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
-
     http://www.apache.org/licenses/LICENSE-2.0
-
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -14,7 +11,9 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package plugins
+package golang
 
-// DefaultNameQualifier is the suffix appended to all kubebuilder plugin names.
-const DefaultNameQualifier = "kubebuilder.io"
+import "sigs.k8s.io/kubebuilder/v3/pkg/plugins"
+
+// DefaultGoNameQualifier is the suffix appended to all kubebuilder plugin names for Golang operators.
+const DefaultGoNameQualifier = "go." + plugins.DefaultNameQualifier
