+++ pkg/plugins/golang/v3/plugin.go
@@ -23,7 +23,7 @@ import (
 	"sigs.k8s.io/kubebuilder/v3/pkg/plugins"
 )
 
-const pluginName = "go" + plugins.DefaultNameQualifier
+const pluginName = "go." + plugins.DefaultNameQualifier
 
 var (
 	pluginVersion            = plugin.Version{Number: 3}
